% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_googlesheets.R
\docType{data}
\name{KMLDocument}
\alias{KMLDocument}
\title{KML Document}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KMLDocument
}
\description{
An R6 class handling all properties of an kml file
}
\section{Usage}{

\preformatted{kml_file <- KMLDocument$new(filename = "test.kml",ignore_altitude=TRUE)}

this calls \code{\link[omanymore]{kml_to_koords}} with the specific parameters and generates the

\preformatted{kml_file$xml_data} list
}

\section{Arguments}{

\describe{
  \item{filename}{A \code{character} describing where the kml file is located.}
   \item{ignore_altitude}{A \code{logical} describing wether the kml altitudes shall be read}
   }
}

\keyword{datasets}
